requirements.txt
arrow==1.1.0
asgiref==3.3.4
autopep8==1.5.7
cake==0.2.3
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
Django==3.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
dlib==19.22.0
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.4
Jinja2==2.11.3
jinja2-time==0.2.0
make==0.1.6.post2
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.20.2
opencv-python==4.5.1.48
openpyxl==3.0.7
Pillow==8.2.0
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Python==3.8.3